+++ command/operator_autopilot.go
@@ -25,7 +25,7 @@ Usage: nomad operator autopilot <subcommand> [options]
   This command groups subcommands for interacting with Nomad's Autopilot
   subsystem. Autopilot provides automatic, operator-friendly management of Nomad
   servers. The command can be used to view or modify the current Autopilot
-  configuration.
+  configuration. For a full guide see: https://www.nomadproject.io/guides/autopilot.html
 
   Get the current Autopilot configuration: